+++ tsconfig.cucumber.json
@@ -1,0 +1,8 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "module": "commonjs",
+        "outDir": "dist-cucumber",
+        "noEmit": true
+    }
+}
