+++ tsconfig.json
@@ -13,7 +13,8 @@
     "experimentalDecorators": true,
     "importHelpers": true,
     "target": "ES2022",
-    "module": "preserve",
+    "module": "esnext",
+    "moduleResolution": "node",
     "esModuleInterop": true,
   },
   "angularCompilerOptions": {
